requirements.txt
# Core Framework
flet==0.21.2

# Integração SharePoint
Office365-REST-Python-Client>=2.5.3

# Processamento de Dados
pandas>=2.0.0
pytz>=2023.3

# Utilitários
python-dateutil>=2.8.2

# Logging e Monitoramento (opcional para GCP)
google-cloud-logging>=3.8.0

# Development (remover em produção se necessário)
python-dotenv>=1.0.0